moved stylesheet into seperate file, resolves issue #23
